CompositeLD.jl Documentation

Overview

CompositeLD is a package using SnpArrays to perform some usefull calculations related LD. In particular :

Calculate LD between two SNPs (given position or id in PLINK files)

Get the LD matrix of a group of SNPs

Get Strongly correlated SNPs to a given group of SNPs

Clumping given SNPs in a prioritized order.

Example

using SnpArrays
using CompositeLD

data = SnpData(SnpArrays.datadir(datapath))

ld = ld_r2("rs1234", "rs654321", data)

ld_mat = getLDmat(data, [(1, 459876), (1, 58735), (2, 97654)])

strong = getStrongLD(data, ["rs1234", "rs5678"])

kept_v_b = clump(data, [(1, 459876), (1, 58735), (2, 97654)])

Contents

CompositeLD.jl Documentation

Overview

Example

Contents

Functions

Index

Functions

CompositeLD.formatSnpData! — Function

format Genotype information contained in SnpData for snp sorted search based on chromosome and position (snpid). Adds a column chrpos (chr::Int8, pos::Int) in snpinfo and sorts snpinfo by chrpos (snpid). Adds a column containing original indices. returns nothing

Examples

ref = SnpData(datadir("some/data"))

formatSnpData!(ref, :chr_pos)

formatSnpData!(ref, :snpid)

Note this function does not take into account the user might want to write SnpData in bed, bim, fam format. The changes done by this function can and will cause problems if writing plink files after calling formatSnpData!.

CompositeLD.ld_r2 — Function

LD r² composite of pair of SNPs given genotype matrix and indices of snps in matrix

Examples :

ref = SnpData(datadir("some/data"))

ld_r2([0x00, 0x01, 0x03], [0x00, 0x00, 0x03])

ld_r2(1, 2, ref.snparray)

formatSnpData!(ref, :snpid)

ld_r2("rs123", "rs456", ref, formated = true)

ld_r2((1, 234), (1, 567), ref)

If formatSnpData has already been called on good snp info type (:chr_pos or :snpid), formated = true option does not verify or modify formating. See formatSnpData!.
returns a Float64

CompositeLD.getLDmat — Function

LD r² composite matrix of n SNPs from index in given SnpArray

Examples :

ref = SnpData(datadir("some/data"))

M::Matrix{Float64} = getLDmat([2, 4, 8, 16], ref.snparray)

getLDmat([(1, 123), (1, 456), (1, 789)], ref)

formatSnpData!(ref, :snpid)

M::Matrix{Float64} = getLDmat(["rs123", "rs456", "rs789"], ref, formated = true)

If formatSnpData has already been called on good snp info type (:chr_pos or :snpid), formated = true option does not verify or modify formating. See formatSnpData!.

Get correlation Matrix for specifies snps (tuple of the form (chr, pos) given reference genotype SnpData.

CompositeLD.getStrongLD — Function

Get strongly correlated variants from list.

arguments :

ref&#95;genotypes::SnpData : reference genotypes (formated by formatSnpData! or not) snps::Union{AbstractVector{<:Tuple{integer, Integer}}, AbstractVector{<:AbstractString}} : snps for which to search proxis

options :

formated::Bool : indicates if ref SnpData is already formated according to :chr_pos (chr, pos) or :snpid (id as string)
window::Integer : maximal distance from snip for ld calculations
r2&#95;tresh::AbstractFloat : minimal r² for 2 snps to be considered strongly correlated.

Examples

ref = SnpData(datadir("some/data"))

l::Vector{String} = getStrongLD([(1, 123), (1, 456), (1, 789)], ref, window = 250000)

formatSnpData!(ref, :snpid)

l::Vector{String} = getStrongLD(["rs123", "rs456", "rs789"], ref, formated = true, r2_tresh = 0.9)

If formatSnpData has already been called on good snp info type (:chr_pos or :snpid), formated = true option does not verify or modify formating. See formatSnpData!.

CompositeLD.clump — Function

Implementation of the clumping algorithm prioritising first snps in given Vector and formated Genotypes SnpData (see formatSnpData!) returns a vector of booean indication if each given snp is kept

arguments :

ref&#95;genotypes::SnpData : reference genotypes (formated by formatSnpData! or not) snps::Union{AbstractVector{<:Tuple{integer, Integer}}, AbstractVector{<:AbstractString}} : snps to clump

options :

formated::Bool : indicates if ref SnpData is already formated according to :chr_pos (chr, pos) or :snpid (id as string)
r2&#95;tresh::AbstractFloat : minimal r² for 2 snps to be considered strongly correlated.

Examples :

julia> ref = SnpData(datadir("some/data"));

julia> kept_v_b::Vector{Bool} = clump([(1, 123), (1, 456), (1, 789)], ref)
3-element Vector{Int64}:
 1
 0
 1

julia> formatSnpData!(ref, :snpid);

julia> kept_v_b::Vector{Bool} = clump(["rs123", "rs456", "rs789"], ref, formated = true)
3-element Vector{Int64}:
 1
 0
 1

If formatSnpData has already been called on good snp info type (:chr_pos or :snpid), formated = true option does not verify or modify formating. See formatSnpData!.

CompositeLD.tclump — Function

threaded implementation of the clumping algorithm prioritising first snps in given Vector and formated Genotypes SnpData (see formatSnpData!) returns a vector of booean indication if each given snp is kept

arguments :

ref&#95;genotypes::SnpData : reference genotypes (formated by formatSnpData! or not) snps::Union{AbstractVector{<:Tuple{integer, Integer}}, AbstractVector{<:AbstractString}} : snps to clump

options :

formated::Bool : indicates if ref SnpData is already formated according to :chr_pos (chr, pos) or :snpid (id as string)
r2&#95;tresh::AbstractFloat : minimal r² for 2 snps to be considered strongly correlated.

Examples :

ref = SnpData(datadir("some/data"))

kept_v_b::Vector{Bool} = tclump([(1, 123), (1, 456), (1, 789)], ref)

formatSnpData!(ref, :snpid)

kept_v_b::Vector{Bool} = tclump(["rs123", "rs456", "rs789"], ref, formated = true)

If formatSnpData has already been called on good snp info type (:chr_pos or :snpid), formated = true option does not verify or modify formating. See formatSnpData!.

Index

CompositeLD.clump

CompositeLD.formatSnpData!

CompositeLD.getLDmat

CompositeLD.getStrongLD

CompositeLD.ld_r2

CompositeLD.tclump